+++ include/swgLib/pass.hpp
@@ -46,7 +46,7 @@ namespace ml
 		std::size_t read(std::istream& file);
 
 	protected:
-		uint8_t _version; // 0000..0010
+		uint32_t _version; // 0000..0010
 
 		bool   _hasPixelShader;
